- add action to set scrollable speed - add animation.set action - add BlendMode to sprite and ImageContainers/TexturePacks - add color to FreeTypeFont - add CompositeSprite - add CompositeTexture - add DC6 image decoder from Worldstone - add DCC image decoder from Worldstone - add drawable bringtofront/sendtoback actions - add horizontalAlign to string Text - add offsets to imagecontainers - add palette to BitmapFontTexturePack - add rapidjson functions to get string and string_view - add support for diablo 2 fonts - add support for diablo 2 palettes - add support to linking without Diablo file format support - always calculate Panel's draw position and size - copy sf::text from SFML and add align code to it - drawable.move - fix BindableText bug - fix video hang with diablo2 videos - increase framerate limit - proper align string text - refactor BitmapFont to use vertex array (faster for big texts) - refactor BitmapText - refactor Image - refactor inputEvents to support keyRelease events and combinations of input events - refactor key press to support repeated key presses - refactor LoadingScreen to extend Animation - refactor player walk code - refactor ScrollableText into generic Scrollable for any UIObject - refactor Sprite2 - refactor text classes - refactor TexturePacks and ImageContainers - remove isIndexed from TexturePack (assume it is, if it has a palette) - rename player.moveToClick to player.walkToClick - set window size to be minimum size if lower than minimum size - store all resources in unordered_multimap - store songs separate from other resources - update A* - update Animation - update BaseAnimation - update gsl - update player animation range to use TexturePacks - update rapidjson with string_view support - update refSize and minSize after window is open - use string_view in ParseUtils - use texturePacks in images/buttons - update gamefiles - add flare player textures for all equipped weapon types in diablo - initial diablo 2 gamefiles (requires DIABDAT) - update flare credits - walk using arrow keys

+++ src/Game/Item.h
@@ -82,7 +82,7 @@ public:
 	void resetDropAnimation(Level& level) noexcept;
 	void resetDropAnimation(LevelMap& map) noexcept;
 
-	virtual bool getTexture(size_t textureNumber, TextureInfo& ti) const;
+	virtual bool getTexture(uint32_t textureNumber, TextureInfo& ti) const;
 
 	virtual bool getNumberProp(const std::string_view prop, Number32& value) const;
 	virtual bool Passable() const noexcept { return true; }
