+++ src/Queryable.h
@@ -39,5 +39,5 @@ public:
 
 	virtual const Queryable* getQueryable(const std::string_view prop) const { return nullptr; }
 
-	virtual bool getTexture(size_t textureNumber, TextureInfo& ti) const { return false; }
+	virtual bool getTexture(uint32_t textureNumber, TextureInfo& ti) const { return false; }
 };
